Disabled BigPOI
BigPOI was causing weird behavior after integrating the getAllPOIs route from the backend.
I've commented it out for now to show that the getAllPOIs route is working correctly - notice that the tabCards populate with Trails/Attractions/Restaurants.
I've also uploaded the uncommented version as tabCards.withBigPOI
Note that tempData that was used previously contained attributes that are no longer available in the latest schema. Thus, the cards don't display all the information correctly. This will need to be corrected.